Fix int ids

+++ src/GitLabApiClient/GroupsClient.cs
@@ -163,7 +163,7 @@ namespace GitLabApiClient
         /// </summary>
         /// <param name="groupId">The ID, path or <see cref="Group"/> of the group.</param>
         /// <param name="milestoneId">Id of the milestone.</param>
-        public async Task<Milestone> GetMilestoneAsync(GroupId groupId, int milestoneId) =>
+        public async Task<Milestone> GetMilestoneAsync(GroupId groupId, long milestoneId) =>
             await _httpFacade.Get<Milestone>($"groups/{groupId}/milestones/{milestoneId}");
 
         /// <summary>
@@ -241,7 +241,8 @@ namespace GitLabApiClient
         /// <param name="milestoneId">The ID of the group's milestone.</param>
         /// <param name="request">Update milestone request.</param>
         /// <returns>Newly modified milestone.</returns>
-        public async Task<Milestone> UpdateMilestoneAsync(GroupId groupId, int milestoneId, UpdateGroupMilestoneRequest request)
+        public async Task<Milestone> UpdateMilestoneAsync(GroupId groupId, long milestoneId,
+            UpdateGroupMilestoneRequest request)
         {
             Guard.NotNull(request, nameof(request));
             return await _httpFacade.Put<Milestone>($"groups/{groupId}/milestones/{milestoneId}", request);
